JAMES-2342 Move MessageMoves to mailbox-api
